backend/requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.2
django-cors-headers==2.4.0
django-extensions==2.1.3
django-localflavor==2.1
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
names==0.3.0
Pillow==5.3.0
python-decouple==3.1
